Moved model traits into proper directory
