+++ app/Models/Traits/CompanyableChildTrait.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Traits;
+
+use App\Models\CompanyableChildScope;
 
 trait CompanyableChildTrait
 {
